% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{evaluate_my_model}
\alias{evaluate_my_model}
\title{Determine existence and uniqueness properties of the New Keynesian Monetary Model}
\usage{
evaluate_my_model(deep_params = NULL, threshold = 1.01,
  my_model = "Lubik_Marzo_26_0", tol = sqrt(.Machine$double.eps),
  verbose = FALSE)
}
\arguments{
\item{deep_params}{List object containing the deep parameters for the respective model.
Default is the model by Lubik and Marzo (2007), equation (26) with \code{i = 0}.
For more detail on the deep parameters see \link{linear_parameters_LM_26_0}.}

\item{threshold}{Numeric value. Generalized eigenvalues larger than "threshold" are considered unstable.}

\item{my_model}{String indicating which model should be evaluated.
Default is the model by Lubik and Marzo (2007), equation (26) with \code{i = 0}.}

\item{tol}{Threshold to determine what is considered practically zero.
Default set to square root of machine precision.}

\item{verbose}{Boolean variable to indicate whether matrices in Sims' canonical form,
the QZ decomposition, and the matrices determining existence and uniqueness should be printed.
Default is set to FALSE.}
}
\value{
Object of type "data_frame". Contains information about the existence and uniqueness properties.
  In particular, the variables are
  \itemize{
    \item \strong{existence_boolean:} TRUE if a stable and causal solution exists
    \item \strong{existence_dim_kernel:} Dimension of the kernel of the matrix pertaining to the endogenous forecast errors in the existence condition.
    \item \strong{uniqueness_boolean:} TRUE if the uniqueness condition is satisfied
    \item \strong{indeterminacy_dim:} Dimension of indeterminacy as characterized in the paper accompanying this package
    \item \strong{indeterminacy_smallest_sv:} Smallest singular value of the intersection of the spaces in the uniqueness condition. The smaller, the "more unique".
  }
}
\description{
Function for evaluating the existence and uniqueness properties of a system in Sims' canonical form.
}
