% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_forms.R
\name{linear_parameters_LM_26_0}
\alias{linear_parameters_LM_26_0}
\title{Obtain linear parameters of model by Lubik and Marzo (2007) equation (26)}
\usage{
linear_parameters_LM_26_0(s = 1, b = 0.99, k = 0.5, f_y = 0.8,
  f_p = 0.7, f_r = 1.2)
}
\arguments{
\item{s}{Elasticity of intertemporal substitution and the inverse of the coefficient of relative risk aversion.}

\item{b}{Subjective time preference rate}

\item{k}{Inverse of elasticity of aggregate supply with respect to inflation}

\item{f_y}{Measures the elasticity of interest rate response w.r.t. output.}

\item{f_p}{Measures the elasticity of interest rate response w.r.t. inflation.}

\item{f_r}{Interest smoothing coefficient (page 27 in LM07)}
}
\value{
List object containing the parameter matrices \eqn{\Gamma_0, \Gamma_1, \Psi, \Pi} as list objects.
}
\description{
This function returns a DSGE model as described in  used in \href{https://doi.org/10.1016/j.iref.2004.09.008}{Lubik Marzo (2007)}.
In particular, it takes the deep parameters as input and returns the matrices occurring in Sims' canonical form.
}
