% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qz_ordered.R
\name{qz_ordered}
\alias{qz_ordered}
\title{Decoupling into Stable and Unstable Systems}
\usage{
qz_ordered(threshold = 1.01, my_Gamma_0, my_Gamma_1,
  tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{threshold}{Threshold to define when a generalized eigenvalue is considered as unstable}

\item{my_Gamma_0}{Matrix from Sims' canonical form}

\item{my_Gamma_1}{Matrix from Sims' canonical form}

\item{tol}{Numeric tolerance level. Default value set to the square root of machine precision.}
}
\value{
List object of the same form as the output of
}
\description{
Uses the \link[QZ]{qz} decomposition of \eqn{\Gamma_0 , \Gamma_1} from Sims' canoncial form to decouple into a stable and an unstable part.
It rearranges them so that all cases of \deqn{ \frac{\Omega_{i,i}}{\Lambda_{i,i}} > stake} are in the bottom right corner.
It ensures that the product \code{t(Q} \eqn{\Lambda} \code{t(Z)} and \code{t(Q)} \eqn{\Omega} \code{t(Z)} remains the same
}
\details{
Note that the function \link[QZ]{qz} returns the QZ decomposition such that
\deqn{( \Gamma_0, \Gamma_1 ) = ( VSL  S  Conj(t(VSR)), VSL * T * Conj(t(VSR)) )}
while in Sims' notation (page 9, formula (32) in \href{Sims (2001)}{https://doi.org/10.1023/A:1020517101123}) the QZ decomposition corresponds to
\deqn{( \Gamma_0, \Gamma_1 ) = ( Conj(t(Q)) \Lambda Conj(t(Z)) , Conj(t(Q)) \Omega Conj(t(Z)) )}.
The code is intended to keep the notation close to Sims (2001) and adjusts the outputs of the \link[QZ]{qz} function accordingly.

Some parts of this function are based on R code by Sims (22 Feb 2004) which in turn is based on earlier matlab code (finished 27 Apr 2000).
Retrieved on 14 June 2017 from \url{http://sims.princeton.edu/yftp/gensys/Rfiles/}.
}
