% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qz_ordered.R
\name{qz_switch}
\alias{qz_switch}
\title{Switch diagonal elements of upper-triangular matrices}
\usage{
qz_switch(my_qz_list, i = 1, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{my_qz_list}{List object containing the orthogonal and upper-triangular matrices returned by the QZ-decomposition in R.}

\item{i}{Index of the diagonal element to be switched with the following one, i.e. \code{i+1}}

\item{tol}{Numeric tolerance level. Default set to square root of machine precision.}
}
\value{
List object
}
\description{
This function switches the diagonal elements of the upper-triangular matrices \eqn{\Lambda} and \eqn{\Omega} in the QZ-decomposition.
It is a helper function for making the QZ-decomposition unique in the case where all generalized eigenvalues are different from each other.
}
\details{
Sims' comment: (Note that the primes might be messed up!)
Takes U.T. matrices a, b, orthonormal matrices q,z, interchanges diagonal elements i and i+1 of both a and b, while maintaining  qaz' and qbz' unchanged.
If diagonal elements of a and b are zero at matching positions, the returned a will have zeros at both positions on the diagonal.
This is natural behavior if this routine is used to drive all zeros on the diagonal of a to the lower right, but in this case the qz transformation is not unique and it is not possible simply to switch the positions of the diagonal elements of both a and b.

See also https://doi.org/10.1007/978-94-015-8196-7_11
}
