% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qz_ordered.R
\name{qz_switch}
\alias{qz_switch}
\title{Switch diagonal elements of upper-triangular matrices}
\usage{
qz_switch(my_qz_list, i = 1, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{my_qz_list}{List object containing the orthogonal and upper-triangular matrices returned by the QZ-decomposition in R.}

\item{i}{Index of the diagonal element to be switched with the following one, i.e. \code{i+1}}

\item{tol}{Numeric tolerance level. Default set to square root of machine precision.}
}
\value{
List object \code{qz_list} with elements
  \itemize{
    \item \strong{Lambda:} Upper-triangular matrix in Sims' notation.
    \item \strong{Omega:} Upper-triangular matrix in Sims' notation.
    \item \strong{Z:} Orthogonal matrix in Sims' notation.
    \item \strong{Q:} Orthogonal matrix in Sims' notation.
    \item \strong{gev_abs:} The absolute values of generalized eigenvalues.
    \item \strong{ev_pairs:} The pairs of diagonal elements of \eqn{\Lambda} and \eqn{\Omega}.
  }
}
\description{
This function switches the diagonal elements of the upper-triangular matrices \eqn{\Lambda} and \eqn{\Omega} in the QZ-decomposition.
It is a helper function for making the QZ-decomposition unique in the case where all generalized eigenvalues are different from each other.
}
\details{
Based on the code by Christopher A. Sims, retrieved on 18 June 2017 from \url{http://sims.princeton.edu/yftp/gensys/Rfiles/}.
See also \href{https://doi.org/10.1007/978-94-015-8196-7_11}{Kagstrom (1993)}
}
